[BREAKING Router] Separates server from client sources.
