+++ src/Artsy/index.tsx
@@ -1,3 +1,2 @@
 export * from "./Analytics"
-export * from "./Router"
 export { ContextProvider, ContextConsumer, ContextProps } from "./SystemContext"
